+++ src/app/utils/range.ts
@@ -1,7 +1,8 @@
 /**
+ * Make range. Return range includes 'from' and 'to' both.
  * @param from
  * @param to
  * @param step
  */
-export const range = (from: number, to: number, step: number) =>
+export const range = (from: number, to: number, step = 1) =>
   [...Array(Math.floor((to - from) / step) + 1)].map((_, i) => from + i * step);
